# openapi.yaml
openapi: 3.1.0
info:
  title: OrbitPay API
  version: 1.0.0
  description: OrbitPay analytics and payment-streaming API
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
...
paths:
...
  /api/vesting:
    get:
      summary: List vesting schedules for a beneficiary
      tags: [Vesting]
      parameters:
        - in: query
          name: beneficiary
          required: true
          schema:
            type: string
            pattern: '^G[A-Z2-7]{55}$'
      responses:
        '200':
          description: Vesting schedules
        '400':
          description: Invalid or missing beneficiary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/vesting/{id}/progress:
    get:
      summary: Get vesting progress for a schedule
      tags: [Vesting]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vesting progress
        '404':
          description: Schedule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'